Created React app
I decided to do this task using React and Axios
I created React app and added a component called students
I made sure I add the component to the App.js one
I added .env file to keep my token and the school id
After that I worked on the fetch requests to the API in order to be able to get some data
I added an axios request with bearer authorization header as that was suggested as best practice in the documentation
I added 3 calls to the API:
 - first one to get the students
 - second one to get the teachers
 - third one to get the classes
It all worked and I rendered those to the screen.
However, I realised that I would not get the students in a particular class that way and I have to think of another method.
I found out that I have to use some of the included properties to do that. I spent a lot of time to work it out as I was not sure how to use includes and could not find much information in the documentation. At the end I managed to get to the data I need by adding include= and whatever data I need from the meta data.
So I believe I will have to get rid of all of the data I have on the FE apart from the teachers names and add a function which will get called on click to request the details of this particular teacher