Got the employees and classes display
I had an issue with the way I created my app yesterday and had to go over it and completely redo it. I was using class components with componentDidMount which made my JSX not working correctly. So I changed the whole code to functional component and managed to get it display those elements.
I displayed them by using map function in the JSX, second one being nested in the first map.
To display the students in the classes I had to create a new axios request which I did and decided to create a function which should be called on click on the class name and it should display the students for this class. However, I was not able to do that yet.