Students data rendered
I have managed to get the students data rendered to the screen
I decided to use another fetch call for that as my previous one did not work. I got rid of the getStudents() function with its API call because I could not make it call a particular class id.
I created another axios function which calls the API and returns students based on the class ID. I used the ID as a function parameter and take it from the user click. That way I manage to call the function and return the students only for the particular class.
After I get the data, I created another function to make the returned data I need into array. After that I remove the commas and display it to the user.
Lastly I added some CSS to make it appear on the screen regardless of the user scrolling