Some last touches
I added some styling and comments to the code.
I have made it work and just one issue is still ongoing - you need to click twice on the class name to render it as my axios request is synchronous. I tried different ways to make the app act differently such as useEffect hook for this function, make it asynchronous with async and await but they all did not work for different reasons.
I will need to read a bit more about useEffect and useState hooks and fetching with axios to see if I will be able to make the function wait for the response before go to the next lines in the snippet.